都道府県をRSCでフェッチ
