+++ src/feature/populationTrends/pages/PopulationTrendsPage.tsx
@@ -6,12 +6,15 @@ import DataSetSelector from "@/feature/populationTrends/components/DataSetSelect
 import PopulationChart from "@/feature/populationTrends/components/PopulationChart";
 import PrefectureSelector from "@/feature/populationTrends/components/PrefectureSelector";
 import ScrollButton from "@/feature/populationTrends/components/ScrollButton";
-import useFetchPrefectures from "@/feature/populationTrends/hooks/useFetchPrefectures";
 import usePopulationData from "@/feature/populationTrends/hooks/usePopulationData";
 import usePrefectureSelection from "@/feature/populationTrends/hooks/usePrefectureSelection";
+import { Prefecture } from "@/types/prefecture";
 
-export default function PopulationTrendsPage() {
-  const { prefectures, loading, error } = useFetchPrefectures();
+interface PopulationTrendsPageProps {
+  prefectures: Prefecture[];
+}
+
+export default function PopulationTrendsPage({ prefectures }: PopulationTrendsPageProps) {
   const { selectedPrefs, togglePrefecture, toggleRegion } = usePrefectureSelection(prefectures);
 
   const [selectedDataSet, setSelectedDataSet] = useState<string>("総人口");
@@ -50,18 +53,9 @@ export default function PopulationTrendsPage() {
       </h1>
       <div className="h-10" />
       <div className="w-full">
-        {error && (
-          <div
-            className="relative rounded border border-red-400 bg-red-100 px-4 py-3 text-red-700"
-            role="alert"
-          >
-            <strong className="font-bold">Error:</strong>
-            <span className="block sm:inline"> {error}</span>
-          </div>
-        )}
         <section id="prefecture">
           <PrefectureSelector
-            loading={loading}
+            loading={false}
             prefectures={prefectures}
             selectedPrefs={selectedPrefs}
             togglePrefecture={togglePrefecture}
